linux_5.10: weekly rls 2026.06.30
7767ea, [fix](eth):Cancel unnecessary network checks and printing
f65486, fix(sdhci): skip PWR_EN pinmux when pin is reused by other device
b5b777, [fix][pwm]: fix 100% duty polarity lost on multi-channel config (CV184X)

+++ include/linux/cost_time.h
@@ -30,11 +30,11 @@ static struct _time_records *time_records __attribute__((unused)) =
 		__tmp;                                                         \
 	})
 
-unsigned int read_count_tick(void);
+uint64_t read_count_tick(void);
 
-unsigned int read_time_ms(void);
+uint64_t read_time_ms(void);
 
-unsigned int read_time_us(void);
+uint64_t read_time_us(void);
 
 void print_cost_time(unsigned int rettime, const char *func);
